Update Storybook version
